Updated check_deployment_existing.sh

+++ EnvironmentSetting/check_deployment_existing.sh
@@ -2,4 +2,9 @@
 deployment_name="goweb-deployment"
 namespace_name="azuredevopsapp"
 kubectl get deployment -n ${namespace_name} | grep -i ${deployment_name}
-kubectl delete deployment ${deployment_name} -n ${namespace_name}
+if[ $? -eq 0 ]
+then
+  kubectl delete deployment ${deployment_name} -n ${namespace_name}
+else
+  echo "No old deployment! start to deploy."
+fi
